requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/prod.in
#
absl-py==1.2.0
    # via tensorboard
aiohttp==3.8.1
    # via fsspec
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==22.1.0
    # via aiohttp
boltons==21.0.0
    # via -r requirements/prod.in
cachetools==5.2.0
    # via google-auth
cassandra-driver==3.25.0
    # via -r requirements/prod.in
certifi==2022.9.14
    # via requests
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   distributed
    #   flask
    #   geomet
cloudpickle==2.2.0
    # via
    #   dask
    #   dask-glm
    #   distributed
dask[array,dataframe]==2022.9.1
    # via
    #   -r requirements/prod.in
    #   dask-glm
    #   dask-ml
    #   distributed
dask-glm==0.2.0
    # via dask-ml
dask-ml==2022.5.27
    # via -r requirements/prod.in
deprecated==1.2.13
    # via redis
distributed==2022.9.1
    # via dask-ml
editdistance==0.6.0
    # via -r requirements/prod.in
flask==2.2.2
    # via -r requirements/prod.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.8.2
    # via
    #   dask
    #   pytorch-lightning
geomet==0.2.1.post1
    # via cassandra-driver
google-auth==2.11.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
graphviz==0.20.1
    # via -r requirements/prod.in
grpcio==1.48.1
    # via tensorboard
h5py==3.7.0
    # via -r requirements/prod.in
heapdict==1.0.1
    # via zict
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==4.12.0
    # via
    #   flask
    #   markdown
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   distributed
    #   flask
joblib==1.2.0
    # via scikit-learn
llvmlite==0.39.1
    # via numba
locket==1.0.0
    # via
    #   distributed
    #   partd
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
msgpack==1.0.4
    # via distributed
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
multipledispatch==0.6.0
    # via
    #   dask-glm
    #   dask-ml
neo4j==5.0.0
    # via -r requirements/prod.in
numba==0.56.2
    # via dask-ml
numpy==1.23.3
    # via
    #   -r requirements/prod.in
    #   dask
    #   dask-ml
    #   h5py
    #   numba
    #   pandas
    #   pytorch-lightning
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   torchmetrics
    #   torchvision
oauthlib==3.2.1
    # via requests-oauthlib
packaging==21.3
    # via
    #   dask
    #   dask-ml
    #   distributed
    #   pytorch-lightning
    #   redis
    #   torchmetrics
pandas==1.4.4
    # via
    #   -r requirements/prod.in
    #   dask
    #   dask-ml
partd==1.3.0
    # via dask
pillow==9.2.0
    # via torchvision
protobuf==3.19.5
    # via tensorboard
psutil==5.9.2
    # via distributed
psycopg2==2.9.3
    # via -r requirements/prod.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydeprecate==0.3.2
    # via pytorch-lightning
pymongo==4.2.0
    # via -r requirements/prod.in
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via pandas
pytorch-lightning==1.7.6
    # via -r requirements/prod.in
pytz==2022.2.1
    # via
    #   neo4j
    #   pandas
pyyaml==6.0
    # via
    #   dask
    #   distributed
    #   pytorch-lightning
redis==4.3.4
    # via -r requirements/prod.in
requests==2.28.1
    # via
    #   -r requirements/prod.in
    #   fsspec
    #   requests-oauthlib
    #   tensorboard
    #   torchvision
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
scikit-learn==1.1.2
    # via
    #   dask-glm
    #   dask-ml
scipy==1.9.1
    # via
    #   dask-glm
    #   dask-ml
    #   scikit-learn
six==1.16.0
    # via
    #   cassandra-driver
    #   geomet
    #   google-auth
    #   grpcio
    #   multipledispatch
    #   python-dateutil
smart-open==6.2.0
    # via -r requirements/prod.in
sortedcontainers==2.4.0
    # via distributed
tblib==1.7.0
    # via distributed
tensorboard==2.10.0
    # via pytorch-lightning
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
threadpoolctl==3.1.0
    # via scikit-learn
toml==0.10.2
    # via -r requirements/prod.in
toolz==0.12.0
    # via
    #   dask
    #   distributed
    #   partd
torch==1.12.1
    # via
    #   -r requirements/prod.in
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==0.9.3
    # via
    #   -r requirements/prod.in
    #   pytorch-lightning
torchvision==0.13.1
    # via -r requirements/prod.in
tornado==6.1
    # via distributed
tqdm==4.64.1
    # via pytorch-lightning
typing-extensions==4.3.0
    # via
    #   pytorch-lightning
    #   torch
    #   torchvision
urllib3==1.26.12
    # via
    #   distributed
    #   requests
werkzeug==2.2.2
    # via
    #   flask
    #   tensorboard
wheel==0.37.1
    # via tensorboard
wrapt==1.14.1
    # via deprecated
yarl==1.8.1
    # via aiohttp
zict==2.2.0
    # via distributed
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
